agent/requirements_core.txt
# 个人传记撰写Agent核心依赖包

# Web框架
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# AI和机器学习
openai==1.3.7
anthropic==0.7.8

# 图像处理
Pillow==10.1.0

# PDF生成
reportlab==4.0.7

# 二维码生成
qrcode[pil]==7.4.2

# 文本处理
jieba==0.42.1

# 数据处理
pandas==2.1.4
numpy==1.25.2

# 配置管理
pyyaml==6.0.1
python-dotenv==1.0.0
pydantic==2.5.0

# HTTP客户端
httpx==0.25.2

# 文件处理
aiofiles==23.2.1 